requirements.txt
ollama==0.3.3
pydantic==2.9.1
regex==2024.9.11
requests==2.32.3
termcolor==2.5.0
llama-cpp-python==3.2.0
groq==1.2.0
transformers==4.46.3
python-dotenv==1.0.1
protobuf==5.2.8
sentencepiece==0.2.0